небольшие правки в апи

+++ api/db/migrations/20260423165739_new_table_tg_offer.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS tg_house_offer (
     id BIGSERIAL PRIMARY KEY NOT NULL,
     title TEXT NOT NULL,
     description TEXT NOT NULL DEFAULT '',
-    price DECIMAL (12, 2) NOT NULL,
+    price INTEGER NOT NULL,
     type TEXT NOT NULL CHECK (type IN ('RENT', 'SELL')),
     owner_id BIGINT NOT NULL REFERENCES tg_user (id)
 );
